[change] バックエンドに送る処理をボタンのコンポーネントに記述
